refactor(frontend): change `_utils` -> `_hooks`
